fastapi_agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.1
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML
openai-whisper==20231117
gTTS==2.4.0
ollama==0.1.6

# RAG
chromadb==0.4.18
langchain==0.0.350
langchain-community==0.0.1
sentence-transformers==2.2.2

# WebRTC
aiortc==1.6.0
aiohttp==3.9.1
websockets==12.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Audio processing
pydub==0.25.1
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

